refactor: Move TypeForm-specific logic to TypeFormDispatcher.
